job1: move Job1 class to j1 folder
